feat(protocol-designer): Allow slot selection when FF enabled
