+++ protocol-designer/src/components/modules/EditModulesCard.js
@@ -5,6 +5,7 @@ import ModuleRow from './ModuleRow'
 import styles from './styles.css'
 
 import type { ModuleType } from '@opentrons/shared-data'
+import { SUPPORTED_MODULE_TYPES } from '../../modules'
 import type { ModulesForEditModulesCard } from '../../step-forms'
 
 type Props = {
@@ -12,17 +13,13 @@ type Props = {
   openEditModuleModal: (type: ModuleType, moduleId?: string) => mixed,
 }
 
-// TODO (ka 2019-10-24): This will likely be resused a lot,
-// Possble candidate for modules/module-data
-const MODULE_TYPES: Array<ModuleType> = ['magdeck', 'tempdeck', 'thermocycler']
-
 export default function EditModulesCard(props: Props) {
   const { modules, openEditModuleModal } = props
 
   return (
     <Card title="Modules">
       <div className={styles.modules_card_content}>
-        {MODULE_TYPES.map((moduleType, i) => {
+        {SUPPORTED_MODULE_TYPES.map((moduleType, i) => {
           const moduleData = modules[moduleType]
           if (moduleData) {
             return (
